Major overhaul of run_test
Fixed up run_test to make it alot more rigerous than it was. Refer to
comments inside for more details. Allows for testing of both pmem and
iopmem.

Also added a simple build script for my system and a fix in the nvme
IOCTL for a problem with meta-data we have yet to resolve.